Move context for breadcrumb, blockquote, banner, accordion
