+++ src/generic/generic-component-banner/data/demo--hero.js
