+++ src/generic/generic-component-banner/data/demo--video.js
