+++ src/generic/generic-component-blockquote/data/demo.js
@@ -1,0 +1,5 @@
+module.exports = {
+  body:
+    'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor.',
+  author: 'Lorem ipsum',
+};
