+++ src/generic/generic-component-breadcrumb/data/demo.js
@@ -1,0 +1,16 @@
+module.exports = {
+  segments: [
+    {
+      href: '../../example.html#',
+      label: 'European Commission',
+    },
+    {
+      href: '../../example.html#',
+      label: 'Priorities',
+    },
+    {
+      href: '../../example.html#',
+      label: 'Jobs, Growth and Investment',
+    },
+  ],
+};
