+++ src/systems/ec/ec-component/ec-component-accordion/package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "@ecl/ec-base": "^0.0.1",
     "@ecl/ec-style-icon": "^0.0.1",
-    "@ecl/ec-utility-colorize": "^0.0.1"
+    "@ecl/ec-utility-colorize": "^0.0.1",
+    "@ecl/generic-component-accordion": "^0.0.1"
   },
   "publishConfig": { "access": "public" },
   "repository": {
