+++ src/systems/ec/ec-component/ec-component-breadcrumb/package.json
@@ -8,7 +8,8 @@
   "dependencies": {
     "@ecl/ec-base": "^0.0.1",
     "@ecl/ec-component-link": "^0.0.1",
-    "@ecl/ec-utility-screen-reader": "^0.0.1"
+    "@ecl/ec-utility-screen-reader": "^0.0.1",
+    "@ecl/generic-component-breadcrumb": "^0.0.1"
   },
   "publishConfig": { "access": "public" },
   "repository": {
