+++ src/systems/eu/eu-component/eu-component-accordion/package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "@ecl/eu-base": "^0.0.1",
     "@ecl/eu-style-icon": "^0.0.1",
-    "@ecl/eu-utility-colorize": "^0.0.1"
+    "@ecl/eu-utility-colorize": "^0.0.1",
+    "@ecl/generic-component-accordion": "^0.0.1"
   },
   "publishConfig": { "access": "public" },
   "repository": {
