+++ src/systems/eu/eu-component/eu-component-blockquote/eu-component-blockquote.config.js
@@ -1,10 +1,8 @@
+const context = require('@ecl/generic-component-blockquote/data/demo');
+
 module.exports = {
   title: 'Blockquotes',
   label: 'Blockquotes',
   status: 'ready',
-  context: {
-    body:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor.',
-    author: 'Lorem ipsum',
-  },
+  context,
 };
