+++ src/systems/eu/eu-component/eu-component-breadcrumb/eu-component-breadcrumb.config.js
@@ -1,23 +1,10 @@
+const context = require('@ecl/generic-component-breadcrumb/data/demo');
+
 module.exports = {
   title: 'Breadcrumbs',
   label: 'Breadcrumbs',
   status: 'ready',
   tags: ['molecule'],
   preview: '@preview-breadcrumbs',
-  context: {
-    segments: [
-      {
-        href: '../../example.html#',
-        label: 'European Commission',
-      },
-      {
-        href: '../../example.html#',
-        label: 'Priorities',
-      },
-      {
-        href: '../../example.html#',
-        label: 'Jobs, Growth and Investment',
-      },
-    ],
-  },
+  context,
 };
