+++ src/systems/eu/eu-component/eu-component-breadcrumb/package.json
@@ -8,7 +8,8 @@
   "dependencies": {
     "@ecl/eu-base": "^0.0.1",
     "@ecl/eu-component-link": "^0.0.1",
-    "@ecl/eu-utility-screen-reader": "^0.0.1"
+    "@ecl/eu-utility-screen-reader": "^0.0.1",
+    "@ecl/generic-component-breadcrumb": "^0.0.1"
   },
   "publishConfig": { "access": "public" },
   "repository": {
